wordcloud/requirements.txt
streamlit
numpy
pandas
pydaisi
Pillow
